add download xlsx

+++ frontend/src/app/experiment/experiment-creator.component.ts
@@ -1,6 +1,6 @@
 import { Component, ViewChild } from '@angular/core';
 import { NgbCollapseModule } from '@ng-bootstrap/ng-bootstrap';
-import { Experiment } from './experiment';
+import { Experiment, ExperimentStatus } from './experiment';
 import { ExperimentService } from './experiment.service';
 import { CommonModule } from '@angular/common';
 import { RouterModule } from '@angular/router';
@@ -34,6 +34,10 @@ export class ExperimentCreatorComponent {
     .subscribe(experiments => this.experiments = experiments);
   }
 
+  getExperimentStatus(status: number): string {
+    return this.experimentService.getExperimentStatus(status);
+  }
+
   add(name: string, model: string, version: string, program: string, configurationTemperature: string ): void {
     name = name.trim();
     if (!name || !model || !version) { return; }
@@ -45,8 +49,10 @@ export class ExperimentCreatorComponent {
   }
 
   delete(experiment: Experiment): void {
-    this.experiments = this.experiments.filter(h => h !== experiment);
-    this.experimentService.deleteExperiment(experiment.id).subscribe();
+      if (confirm(`Are you sure that you want to delete the experiment ${experiment.name}?`)) {
+        this.experiments = this.experiments.filter(exp => exp !== experiment);
+        this.experimentService.deleteExperiment(experiment.id).subscribe();
+    }
   }
 
 }
